refactor: 데이터베이스 통합
- 하나의 데이터베이스에 장바구니 테이블과 최근 조회한 상품 테이블을 갖습니다
